@@ -7,13 +7,11 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 
-import static java.util.stream.Collectors.toList;
-
 public interface Repository {
     default void storeArtifactPaths(Instance instance, List<Path> files) throws IOException {
         storeArtifacts(instance, files.stream()
                 .map(path -> new Artifact(path.getFileName().toString(), path))
-                .collect(toList()));
+                .toList());
     }
 
     void storeArtifacts(Instance instance, List<Artifact> files) throws IOException;
